refactor: create separate files for various functions
