+++ src/metadata/metadataInterface.ts
@@ -1,0 +1,8 @@
+'use strict';
+
+export interface Metadata {
+  directoryName: string;
+  metaSuffix: string;
+  xmlElement: string;
+  fieldNames: string;
+}
