+++ src/types/xmlParserOptions.ts
@@ -1,0 +1,12 @@
+'use strict';
+
+export const XML_PARSER_OPTION = {
+  commentPropName: '#comment',
+  ignoreAttributes: false,
+  ignoreNameSpace: false,
+  parseTagValue: false,
+  parseNodeValue: false,
+  parseAttributeValue: false,
+  trimValues: true,
+  processEntities: false,
+};
